+++ packages/fcl/src/current-user/exec-service/index.js
@@ -1,13 +1,14 @@
 import {execHttpPost} from "./strategies/http-post"
 import {execIframeRPC} from "./strategies/iframe-rpc"
 import {execPopRPC} from "./strategies/pop-rpc"
+import {execTabRPC} from "./strategies/tab-rpc"
 
 const STRATEGIES = {
   "HTTP/RPC": execHttpPost,
   "HTTP/POST": execHttpPost,
   "IFRAME/RPC": execIframeRPC,
   "POP/RPC": execPopRPC,
-  "TAB/RPC": execPopRPC,
+  "TAB/RPC": execTabRPC,
 }
 
 export async function execService({service, msg = {}, opts = {}}) {