+++ packages/fcl/src/current-user/exec-service/strategies/utils/render-frame.js
@@ -37,5 +37,5 @@ export function renderFrame(src) {
     }
   }
 
-  return [$frame, unmount]
+  return [$frame.contentWindow, unmount]
 }